Remove the druidry hatchet; use PrimalCore
Druidry's hatchet is more-or-less the same as the PrimalCore hatchet. Since this script makes it possible to repair the PrimalCore hatchet using sticks and branches, there is little need to have two such hatchets.

I also went ahead and made the axe able to be repaired similarly, but unlike the hatchet, the axe should retain its enchantments when repaired. It will also start with the Unbreaking 1 enchantment (this'll make it really easy to test if the repair function works properly).