Adding most ores to the bloomery
Granted, historically, the bloomery was only used to process iron, and it did not melt the metal, as iron had a higher melting point than the bloomery was capable of. The bloomery worked by melting the slag and impurities in the iron, leaving an iron "bloom" that needed to be pounded to push the molten slag out of the iron bloom. This produced wrought iron, called that because it had to be worked. Wrought iron had very little carbon, typically less than 0.1%.