Revert input stack change for refractory kiln/oven
I increased the inherited recipe speed, which should make the machines faster and more fuel efficient. I intend on playing with this more.
I am attempting to make the refractory oven and kiln to be on par with the vanilla furnace in terms of fuel efficiency. With processing speed, I want the refractory oven and kiln to overall be faster than the vanilla furnace when you fill the input slots.